+++ tests/test_twopoint_difference.py
@@ -1,4 +1,5 @@
 import numpy as np
+from astropy.io import fits
 import pytest
 
 from stcal.jump.twopoint_difference import calc_med_first_diffs, find_crs
@@ -8,10 +9,8 @@ DQFLAGS = {"JUMP_DET": 4, "SATURATED": 2, "DO_NOT_USE": 1}
 
 @pytest.fixture()
 def setup_cube():
-    def _cube(ngroups, readnoise=10):
-        nints = 1
-        nrows = 204
-        ncols = 204
+
+    def _cube(ngroups, nints=1, nrows=204, ncols=204, readnoise=10):
         rej_threshold = 3
         nframes = 1
         data = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.float32)
@@ -21,6 +20,40 @@ def setup_cube():
         return data, gdq, nframes, read_noise, rej_threshold
 
     return _cube
+
+
+def test_varying_groups(setup_cube):
+    ngroups = 5
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, nints=1, nrows=2, ncols=2, readnoise=8)
+    data[0, :, 0, 0] = [10, 20, 30, 530, 540]
+    data[0, :, 0, 1] = [10, 20, 30, 530, np.nan]
+    data[0, :, 1, 0] = [10, 20, 530, np.nan, np.nan]
+    data[0, :, 1, 1] = [10, 520, np.nan, np.nan, np.nan]
+    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
+        data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
+    )
+    assert np.array_equal(out_gdq[0, :, 0, 0], [0, 0, 0, 4, 0])
+    assert np.array_equal(out_gdq[0, :, 0, 1], [0, 0, 0, 4, 0])
+    assert np.array_equal(out_gdq[0, :, 1, 0], [0, 0, 4, 0, 0])
+    assert np.array_equal(out_gdq[0, :, 1, 1], [0, 0, 0, 0, 0])
+
+
+def test_multint_pixel(setup_cube):
+    ngroups=4
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, nints=7, nrows=2, ncols=2, readnoise=8)
+    data[0, :, 0, 0] = (-24,   -15,     0,    13)
+    data[1, :, 0, 0] = (-24,   -11,     6,    21)
+    data[2, :, 0, 0] = (-40,   -28,   -24,    -4)
+    data[3, :, 0, 0] = (-11,     3,    11,    24)
+    data[4, :, 0, 0] = (-43 ,  -24,   -12,     1)
+    data[5, :, 0, 0] = (-45,  8537, 17380, 17437)
+    data[6, :, 0, 0] = (-178,  -156,  -139,  -125)
+    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
+        data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
+    )
+    assert np.max(out_gdq) == 4  # a CR was found
+    assert (np.array_equal([0, 4, 4, 4], out_gdq[5, :, 0, 0]))
+
 
 
 def test_nocrs_noflux(setup_cube):
@@ -44,20 +77,25 @@ def test_5grps_cr3_noflux(setup_cube):
     )
     assert np.max(out_gdq) == 4  # a CR was found
     assert np.argmax(out_gdq[0, :, 100, 100]) == 2  # find the CR in the expected group
-
-
-def test_5grps_cr2_noflux(setup_cube):
-    ngroups = 5
-    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups)
-
     data[0, 0, 100, 100] = 10.0
     data[0, 1:6, 100, 100] = 1000
+
+def test_4grps_2ints_cr2_noflux(setup_cube):
+    ngroups = 5
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, nints=2, ncols=2, nrows=2)
+    data[0, 1, 1, 1] = 5
+    data[1, 0, 1, 1] = 10.0
+    data[1, 1:6, 1, 1] = 1000
+    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(data, gdq, read_noise, rej_threshold,
+                                                     rej_threshold, rej_threshold, nframes,
+                                                     False, 200, 10, DQFLAGS)
+    assert(4 == np.max(out_gdq))  # a CR was found
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
     )
     assert np.max(out_gdq) == 4  # a CR was found
-    assert np.argmax(out_gdq[0, :, 100, 100]) == 1  # find the CR in the expected group
-
+    assert np.argmax(out_gdq[1, :, 1, 1]) == 1  # find the CR in the expected group
+    assert(1 == np.argmax(out_gdq[1, :, 1, 1]))  # find the CR in the expected group
 
 def test_6grps_negative_differences_zeromedian(setup_cube):
     ngroups = 6
@@ -93,13 +131,31 @@ def test_3grps_cr2_noflux(setup_cube):
     data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups)
     data[0, 0, 100, 100] = 10.0
     data[0, 1:4, 100, 100] = 1000
+    data[0, 0, 99, 99] = 10.0
+    data[0, 2:4, 99, 99] = 1000
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
     )
     assert np.max(out_gdq) == 4  # a CR was found
     #    assert(1,np.argmax(out_gdq[0, :, 100, 100]))  # find the CR in the expected group
-    assert np.array_equal([0, 4, 0], out_gdq[0, :, 100, 100])
+    assert(np.array_equal([0, 4, 0], out_gdq[0, :, 100, 100]))
+    assert (np.array_equal([0, 0, 4], out_gdq[0, :, 99, 99]))
 
+def test_2ints_2grps_noflux(setup_cube):
+    ngroups = 2
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, nints=2, ncols=2, nrows=2)
+    data[0, 0, 1, 1] = 10.0
+    data[0, 1:3, 1, 1] = 1000
+    data[1, 0, 0, 0] = 10.0
+    data[1, 1:3, 0, 0] = 1000
+    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(data, gdq, read_noise, rej_threshold,
+                                                     rej_threshold, rej_threshold, nframes,
+                                                     False, 200, 10, DQFLAGS,
+                                                                         minimum_groups=2)
+ #   assert(4 == np.max(out_gdq))  # a CR was found
+    #    assert(1,np.argmax(out_gdq[0, :, 100, 100]))  # find the CR in the expected group
+    assert(np.array_equal([0, 4], out_gdq[0, :, 1, 1]))
+    assert (np.array_equal([0, 4], out_gdq[1, :, 0, 0]))
 
 def test_4grps_cr2_noflux(setup_cube):
     ngroups = 4
@@ -113,20 +169,24 @@ def test_4grps_cr2_noflux(setup_cube):
     assert np.argmax(out_gdq[0, :, 100, 100]) == 1  # find the CR in the expected group
 
 
-def test_5grps_cr2_nframe2(setup_cube):
-    ngroups = 5
-    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups)
+def test_6grps_cr2_nframe2(setup_cube):
+    ngroups = 6
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, nrows=2, ncols=2)
     nframes = 2
-    data[0, 0, 100, 100] = 10.0
-    data[0, 1, 100, 100] = 500
-    data[0, 2, 100, 100] = 1002
-    data[0, 3, 100, 100] = 1001
-    data[0, 4, 100, 100] = 1005
+    data[0, 0, 1, 1] = 10.0
+    data[0, 1, 1, 1] = 500
+    data[0, 2, 1, 1] = 1002
+    data[0, 3, 1, 1] = 1001
+    data[0, 4, 1, 1] = 1005
+    data[0, 5, 1, 1] = 1015
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
     )
     assert np.max(out_gdq) == 4  # a CR was found
-    assert np.array_equal([0, 4, 4, 0, 0], out_gdq[0, :, 100, 100])
+    assert (np.array_equal([0, 4, 4, 0, 0, 0], out_gdq[0, :, 1, 1]))
+    assert (np.max(out_gdq[0, :, 0, 0]) == 0)
+    assert (np.max(out_gdq[0, :, 1, 0]) == 0)
+    assert (np.max(out_gdq[0, :, 0, 1]) == 0)
 
 
 def test_4grps_twocrs_2nd_4th(setup_cube):
@@ -164,9 +224,9 @@ def test_5grps_twocrs_2nd_5thbig(setup_cube):
     data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups)
     nframes = 1
     data[0, 0, 100, 100] = 10.0
-    data[0, 1, 100, 100] = 60
-    data[0, 2, 100, 100] = 60
-    data[0, 3, 100, 100] = 60
+    data[0, 1, 100, 100] = 600
+    data[0, 2, 100, 100] = 600
+    data[0, 3, 100, 100] = 600
     data[0, 4, 100, 100] = 2115
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
@@ -295,16 +355,16 @@ def test_5grps_cr2_negslope(setup_cube):
     ngroups = 5
     data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups)
     nframes = 1
-    data[0, 0, 100, 100] = 100.0
-    data[0, 1, 100, 100] = 0
-    data[0, 2, 100, 100] = -200
-    data[0, 3, 100, 100] = -260
-    data[0, 4, 100, 100] = -360
+    data[0, 0, 1, 1] = 100.0
+    data[0, 1, 1, 1] = 0
+    data[0, 2, 1, 1] = -200
+    data[0, 3, 1, 1] = -260
+    data[0, 4, 1, 1] = -360
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
     )
     assert np.max(out_gdq) == 4  # a CR was found
-    assert np.array_equal([0, 0, 4, 0, 0], out_gdq[0, :, 100, 100])
+    assert np.array_equal([0, 0, 4, 0, 0], out_gdq[0, :, 1, 1])
 
 
 def test_6grps_1cr(setup_cube):
@@ -360,22 +420,22 @@ def test_8grps_1cr(setup_cube):
 
 def test_9grps_1cr_1sat(setup_cube):
     ngroups = 9
-    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, readnoise=10)
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, readnoise=10, nrows=2, ncols=2)
     nframes = 1
-    data[0, 0, 100, 100] = 0
-    data[0, 1, 100, 100] = 10
-    data[0, 2, 100, 100] = 21
-    data[0, 3, 100, 100] = 33
-    data[0, 4, 100, 100] = 46
-    data[0, 5, 100, 100] = 60
-    data[0, 6, 100, 100] = 1160
-    data[0, 7, 100, 100] = 1175
-    data[0, 8, 100, 100] = 6175
-    gdq[0, 8, 100, 100] = DQFLAGS["SATURATED"]
+    data[0, 0, 1, 1] = 0
+    data[0, 1, 1, 1] = 10
+    data[0, 2, 1, 1] = 21
+    data[0, 3, 1, 1] = 33
+    data[0, 4, 1, 1] = 46
+    data[0, 5, 1, 1] = 60
+    data[0, 6, 1, 1] = 1160
+    data[0, 7, 1, 1] = 1175
+    data[0, 8, 1, 1] = 6175
+    gdq[0, 8, 1, 1] = DQFLAGS["SATURATED"]
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
     )
-    assert out_gdq[0, 6, 100, 100] == 4
+    assert out_gdq[0, 6, 1, 1] == 4
 
 
 def test_10grps_1cr_2sat(setup_cube):
@@ -402,26 +462,26 @@ def test_10grps_1cr_2sat(setup_cube):
 
 def test_11grps_1cr_3sat(setup_cube):
     ngroups = 11
-    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, readnoise=10)
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, readnoise=10, nrows=2, ncols=2)
     nframes = 1
-    data[0, 0, 100, 100] = 0
-    data[0, 1, 100, 100] = 20
-    data[0, 2, 100, 100] = 39
-    data[0, 3, 100, 100] = 57
-    data[0, 4, 100, 100] = 74
-    data[0, 5, 100, 100] = 90
-    data[0, 6, 100, 100] = 1160
-    data[0, 7, 100, 100] = 1175
-    data[0, 8, 100, 100] = 6175
-    data[0, 9, 100, 100] = 6175
-    data[0, 10, 100, 100] = 6175
-    gdq[0, 8, 100, 100] = DQFLAGS["SATURATED"]
-    gdq[0, 9, 100, 100] = DQFLAGS["SATURATED"]
-    gdq[0, 10, 100, 100] = DQFLAGS["SATURATED"]
+    data[0, 0, 1, 1] = 0
+    data[0, 1, 1, 1] = 20
+    data[0, 2, 1, 1] = 39
+    data[0, 3, 1, 1] = 57
+    data[0, 4, 1, 1] = 74
+    data[0, 5, 1, 1] = 90
+    data[0, 6, 1, 1] = 1160
+    data[0, 7, 1, 1] = 1175
+    data[0, 8, 1, 1] = 6175
+    data[0, 9, 1, 1] = 6175
+    data[0, 10, 1, 1] = 6175
+    gdq[0, 8, 1, 1] = DQFLAGS["SATURATED"]
+    gdq[0, 9, 1, 1] = DQFLAGS["SATURATED"]
+    gdq[0, 10, 1, 1] = DQFLAGS["SATURATED"]
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
     )
-    assert out_gdq[0, 6, 100, 100] == 4
+    assert out_gdq[0, 6, 1, 1] == 4
 
 
 def test_11grps_0cr_3donotuse(setup_cube):
@@ -564,23 +624,34 @@ def test_10grps_nocr_2pixels_sigma0(setup_cube):
 
 def test_5grps_satat4_crat3(setup_cube):
     ngroups = 5
-    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, readnoise=5 * np.sqrt(2))
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, nrows=2, ncols=2, readnoise=5 * np.sqrt(2))
     nframes = 1
-    data[0, 0, 100, 100] = 10000
-    data[0, 1, 100, 100] = 30000
-    data[0, 2, 100, 100] = 60000
-    data[0, 3, 100, 100] = 61000
-    data[0, 4, 100, 100] = 61000
-    gdq[0, 3, 100, 100] = DQFLAGS["SATURATED"]
-    gdq[0, 4, 100, 100] = DQFLAGS["SATURATED"]
-    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
-        data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
-    )
-    # assert(4 == np.max(out_gdq))  # no CR was found
-    assert np.array_equal(
-        [0, 0, DQFLAGS["JUMP_DET"], DQFLAGS["SATURATED"], DQFLAGS["SATURATED"]], out_gdq[0, :, 100, 100]
-    )
+    data[0, 0, 1, 1] = 10000
+    data[0, 1, 1, 1] = 20000
+    data[0, 2, 1, 1] = 60000
+    data[0, 3, 1, 1] = 61000
+    data[0, 4, 1, 1] = 61000
+    gdq[0, 3, 1, 1] = DQFLAGS['SATURATED']
+    gdq[0, 4, 1, 1] = DQFLAGS['SATURATED']
 
+    data[0, 0, 0, 1] = 59800
+    data[0, 1, 0, 1] = 59900
+    data[0, 2, 0, 1] = 60000
+    data[0, 3, 0, 1] = 61000
+    data[0, 4, 0, 1] = 61000
+    gdq[0, 3, 0, 1] = DQFLAGS['SATURATED']
+    gdq[0, 4, 0, 1] = DQFLAGS['SATURATED']
+    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(data, gdq, read_noise, rej_threshold,
+                                                     rej_threshold, rej_threshold, nframes,
+                                                     False, 200, 10, DQFLAGS)
+    # assert(4 == np.max(out_gdq))  # no CR was found
+    result = out_gdq[0, :, 1, 1]
+    assert np.array_equal(
+        [0, 0, DQFLAGS['JUMP_DET'], DQFLAGS['SATURATED'], DQFLAGS['SATURATED']],
+        out_gdq[0, :, 1, 1])
+    assert np.array_equal(
+        [0, 0, 0, DQFLAGS['SATURATED'], DQFLAGS['SATURATED']],
+        out_gdq[0, :, 0, 1])
 
 def test_6grps_satat6_crat1(setup_cube):
     ngroups = 6
@@ -672,43 +743,23 @@ def test_10grps_satat8_crsat3and6(setup_cube):
 def test_median_with_saturation(setup_cube):
     ngroups = 10
     # crmag = 1000
-    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, readnoise=5 * np.sqrt(2))
+    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, nrows=2, ncols=2, readnoise=5 * np.sqrt(2))
     nframes = 1
-    data[0, 0, 100, 100] = 0
-    data[0, 1, 100, 100] = 4500
-    data[0, 2, 100, 100] = 9100
-    data[0, 3, 100, 100] = 13800
-    data[0, 4, 100, 100] = 18600
-    data[0, 5, 100, 100] = 40000  # CR
-    data[0, 6, 100, 100] = 44850
-    data[0, 7, 100, 100] = 49900
-    data[0, 8:10, 100, 100] = 60000
-    gdq[0, 7:10, 100, 100] = DQFLAGS["SATURATED"]
-    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
-        data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
-    )
-    assert np.array_equal([0, 0, 0, 0, 0, 4, 0, 2, 2, 2], out_gdq[0, :, 100, 100])
-
-
-def test_median_with_saturation_even_num_sat_frames(setup_cube):
-    ngroups = 10
-    # crmag = 1000
-    data, gdq, nframes, read_noise, rej_threshold = setup_cube(ngroups, readnoise=5 * np.sqrt(2))
-    nframes = 1
-    data[0, 0, 100, 100] = 0
-    data[0, 1, 100, 100] = 4500
-    data[0, 2, 100, 100] = 9100
-    data[0, 3, 100, 100] = 13800
-    data[0, 4, 100, 100] = 18600
-    data[0, 5, 100, 100] = 40000  # CR
-    data[0, 6, 100, 100] = 44850
-    data[0, 7, 100, 100] = 49900
-    data[0, 8:10, 100, 100] = 60000
-    gdq[0, 6:10, 100, 100] = DQFLAGS["SATURATED"]
-    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
-        data, gdq, read_noise, rej_threshold, rej_threshold, rej_threshold, nframes, False, 200, 10, DQFLAGS
-    )
-    assert np.array_equal([0, 0, 0, 0, 0, 4, 2, 2, 2, 2], out_gdq[0, :, 100, 100])
+    data[0, 0, 1, 1] = 0
+    data[0, 1, 1, 1] = 4700
+    data[0, 2, 1, 1] = 9400
+    data[0, 3, 1, 1] = 14300
+    data[0, 4, 1, 1] = 19100
+    data[0, 5, 1, 1] = 40000  # CR
+    data[0, 6, 1, 1] = 44850
+    data[0, 7, 1, 1] = 49900
+    data[0, 8:10, 1, 1] = 60000
+    gdq[0, 7:10, 1, 1] = DQFLAGS['SATURATED']
+    out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(data, gdq, read_noise, rej_threshold,
+                                                     rej_threshold, rej_threshold, nframes,
+                                                     False, 200, 10, DQFLAGS)
+    gdq_value = out_gdq[0, :, 1, 1]
+    assert (np.array_equal([0, 0, 0, 0, 0, 4, 0, 2, 2, 2], out_gdq[0, :, 1, 1]))
 
 
 def test_median_with_saturation_odd_number_final_difference(setup_cube):
@@ -855,7 +906,7 @@ def test_10grps_1cr_afterjump(setup_cube):
     data[0, 8, 100, 100] = 1190
     data[0, 9, 100, 100] = 1209
 
-    after_jump_flag_e1 = 1.0
+    after_jump_flag_e1 = 0.0
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data,
         gdq,
@@ -891,7 +942,7 @@ def test_10grps_1cr_afterjump_2group(setup_cube):
     data[0, 8, 100, 100] = 1190
     data[0, 9, 100, 100] = 1209
 
-    after_jump_flag_e1 = 1.0
+    after_jump_flag_e1 = 0.0
     out_gdq, row_below_gdq, rows_above_gdq, total_crs, stddev = find_crs(
         data,
         gdq,
@@ -1009,15 +1060,17 @@ def test_median_func():
     arr = np.array([1.0, 2.0, 3.0, 4.0, 5])
     assert calc_med_first_diffs(arr) == 2.5
     # 3d array, no nans
-    arr = np.zeros(5 * 2 * 2).reshape(5, 2, 2)
-    arr[:, 0, 0] = np.array([1.0, 2.0, 3.0, 4.0, 5])
-    assert calc_med_first_diffs(arr)[0, 0] == 2.5
+    arr = np.zeros(5 * 2 * 2).reshape(1, 5, 2, 2)
+    arr[0, :, 0, 0] = np.array([1.0, 2.0, 3.0, 5.0, 4])
+    arr[0, 1, 0, 1] = 1.0
+    result = calc_med_first_diffs(arr)
+    assert result[0, 0] == 2.5
     # 1d, with nans
     arr = np.array([1.0, 2.0, 3.0, np.nan, 4.0, 5, np.nan])
     assert calc_med_first_diffs(arr) == 2.5
     # 3d, with nans
-    arr = np.zeros(7 * 2 * 2).reshape(7, 2, 2)
-    arr[:, 0, 0] = np.array([1.0, 2.0, 3.0, np.nan, 4.0, 5, np.nan])
+    arr = np.zeros(7 * 2 * 2).reshape(1, 7, 2, 2)
+    arr[0, :, 0, 0] = np.array([1.0, 2.0, 3.0, np.nan, 4.0, 5, np.nan])
     assert calc_med_first_diffs(arr)[0, 0] == 2.5
 
     # single pix with exactly 4 good diffs, should also clip 1 pix and return median
@@ -1025,43 +1078,119 @@ def test_median_func():
     arr = np.array([1.0, 2.0, 3.0, 4.0])
     assert calc_med_first_diffs(arr) == 2
     # 3d array, no nans
-    arr = np.zeros(4 * 2 * 2).reshape(4, 2, 2)
-    arr[:, 0, 0] = np.array([1.0, 2.0, 3.0, 4.0])
+    arr = np.zeros(4 * 2 * 2).reshape(1, 4, 2, 2)
+    arr[0, :, 0, 0] = np.array([1.0, 2.0, 3.0, 4.0])
     assert calc_med_first_diffs(arr)[0, 0] == 2
     # 1d, with nans
     arr = np.array([1.0, 2.0, 3.0, np.nan, 4.0, np.nan])
     assert calc_med_first_diffs(arr) == 2
     # 3d, with nans
-    arr = np.zeros(6 * 2 * 2).reshape(6, 2, 2)
-    arr[:, 0, 0] = np.array([1.0, 2.0, 3.0, np.nan, 4.0, np.nan])
+    arr = np.zeros(6 * 2 * 2).reshape(1, 6, 2, 2)
+    arr[0, :, 0, 0] = np.array([1.0, 2.0, 3.0, np.nan, 4.0, np.nan])
     assert calc_med_first_diffs(arr)[0, 0] == 2
 
     # single pix with exactly 3 good diffs, should compute median without clipping
     arr = np.array([1.0, 2.0, 3.0])
     assert calc_med_first_diffs(arr) == 2
     # 3d array, no nans
-    arr = np.zeros(3 * 2 * 2).reshape(3, 2, 2)
-    arr[:, 0, 0] = np.array([1.0, 2.0, 3.0])
+    arr = np.zeros(3 * 2 * 2).reshape(1, 3, 2, 2)
+    arr[0, :, 0, 0] = np.array([1.0, 2.0, 3.0])
     assert calc_med_first_diffs(arr)[0, 0] == 2
     # 1d, with nans
     arr = np.array([1.0, 2.0, 3.0, np.nan, np.nan])
     assert calc_med_first_diffs(arr) == 2
     # 3d, with nans
-    arr = np.zeros(5 * 2 * 2).reshape(5, 2, 2)
-    arr[:, 0, 0] = np.array([1.0, 2.0, 3.0, np.nan, np.nan])
+    arr = np.zeros(5 * 2 * 2).reshape(1, 5, 2, 2)
+    arr[0, :, 0, 0] = np.array([1.0, 2.0, 3.0, np.nan, np.nan])
     assert calc_med_first_diffs(arr)[0, 0] == 2
 
     # # single pix with exactly 2 good diffs, should return the element with the minimum abs val
     arr = np.array([-1.0, -2.0])
     assert calc_med_first_diffs(arr) == -1
     # 3d array, no nans
-    arr = np.zeros(2 * 2 * 2).reshape(2, 2, 2)
-    arr[:, 0, 0] = np.array([-1.0, -2.0])
+    arr = np.zeros(2 * 2 * 2).reshape(1, 2, 2, 2)
+    arr[0, :, 0, 0] = np.array([-1.0, -2.0])
     assert calc_med_first_diffs(arr)[0, 0] == -1
     # 1d, with nans
     arr = np.array([-1.0, -2.0, np.nan, np.nan])
     assert calc_med_first_diffs(arr) == -1
     # 3d, with nans
-    arr = np.zeros(4 * 2 * 2).reshape(4, 2, 2)
-    arr[:, 0, 0] = np.array([-1.0, -2.0, np.nan, np.nan])
+    arr = np.zeros(4 * 2 * 2).reshape(1, 4, 2, 2)
+    arr[0, :, 0, 0] = np.array([-1.0, -2.0, np.nan, np.nan])
     assert calc_med_first_diffs(arr)[0, 0] == -1
+@pytest.mark.skip("Used for local testing")
+def test_sigma_clip():
+    hdul = fits.open('lrs_TSOjump_sigmaclip5_00_refpix.fits')
+    data = hdul['SCI'].data * 4.0
+    gdq = hdul['GROUPDQ'].data
+    indata = data[:53, :, :, :]
+    ingdq = gdq[:53, :, :, :]
+    read_noise = np.ones(shape=(indata.shape[2], indata.shape[3]), dtype=np.float32) * 5.9 * 4.0
+    hdul.close()
+    gdq, row_below_gdq, row_above_gdq, total_primary_crs, stddev = find_crs(indata, ingdq, read_noise, 3,
+             4, 5, 1,
+             False, 1000,
+             10, DQFLAGS,
+             after_jump_flag_e1=0.0,
+             after_jump_flag_n1=0,
+             after_jump_flag_e2=0.0,
+             after_jump_flag_n2=0,
+             copy_arrs=True, minimum_groups=3, minimum_sigclip_groups=50,)
+    fits.writeto("outgdq.fits", gdq, overwrite=True)
+@pytest.mark.skip("Used for local testing")
+def test_first_grp_flag_issue():
+    nints = 8
+    nrows = 2
+    ncols = 2
+    ngroups = 10
+    readnoise = 2
+    data = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.float32)
+    data = np.random.normal(0, readnoise, size=(nints, ngroups, nrows, ncols))
+    read_noise = np.full((nrows, ncols), readnoise, dtype=np.float32)
+    gdq = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.uint32)
+
+    gdq[:, 0, :, :] = DQFLAGS['DO_NOT_USE']
+    gdq[1:, 1, :, :] = DQFLAGS['DO_NOT_USE']
+    gdq[:, -1, :, :] = DQFLAGS['DO_NOT_USE']
+    gdq, row_below_gdq, row_above_gdq, total_total_crs, stddev = \
+        find_crs(data, gdq, read_noise, 3, 4, 5, 1, False, 1000, 10, DQFLAGS,
+                 after_jump_flag_e1=0.0, after_jump_flag_n1=0,
+                 after_jump_flag_e2=0.0, after_jump_flag_n2=0,
+                 copy_arrs=True, minimum_groups=3, minimum_sigclip_groups=50)
+    fits.writeto("outgdq.fits",gdq, overwrite=True)
+@pytest.mark.skip("Used for local testing")
+def test_5grp_TSO():
+    nints=20
+    nrows = 2
+    ncols = 2
+    ngroups = 5
+    readnoise = 25
+    data = np.random.normal(0, 0.1 * readnoise, size=(nints, ngroups, nrows, ncols))
+    read_noise = np.full((nrows, ncols), readnoise, dtype=np.float32)
+    gdq = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.uint32)
+    np.expand_dims(gdq, axis=0)
+    np.expand_dims(data, axis=0)
+    gdq[:, 0, :, :] = DQFLAGS['DO_NOT_USE']
+    gdq[:, -1, :, :] = DQFLAGS['DO_NOT_USE']
+    data[0, :, 0, 0] = [21500, 37600, 52082, 65068, 58627]
+    data[0, :, 0, 1] = [21500, 37600, 52082, 65068, 58627]
+    gdq, row_below_gdq, row_above_gdq, total_primary_crs, stddev = \
+        find_crs(data, gdq, read_noise, 3, 4, 5, 1, False, 1000, 10, DQFLAGS,
+                 after_jump_flag_e1=0.0, after_jump_flag_n1=0,
+                 after_jump_flag_e2=0.0, after_jump_flag_n2=0,
+                 copy_arrs=True, minimum_groups=3, minimum_sigclip_groups=5000)
+    fits.writeto("new_gdq.fits", gdq, overwrite=True)
+@pytest.mark.skip("Used for local testing")
+def test_5grp_realTSO():
+    hdul = fits.open("obs2508_cutout_jump.fits")
+    gdq = hdul['groupdq'].data
+    data = hdul['sci'].data
+    readnoise = 25
+    read_noise = np.full((3, 3), readnoise, dtype=np.float32)
+
+    gdq, row_below_gdq, row_above_gdq, total_total_crs, stddev = \
+        find_crs(data, gdq, read_noise, 3, 4, 5, 1, False, 1000, 10, DQFLAGS,
+                 after_jump_flag_e1=0.0, after_jump_flag_n1=0,
+                 after_jump_flag_e2=0.0, after_jump_flag_n2=0,
+                 copy_arrs=True, minimum_groups=3, minimum_sigclip_groups=15000)
+    fits.writeto("new_gdq_cutout.fits", gdq, overwrite=True)